Add support for daily double
In the json file, add
    "double": 1,
to the answer(s) that should be a daily double.

When the tile for such an answer is clicked, you'll be asked how many
points you want to risk.

Note that the value of daily double can be entered freely (without upper
or lower limit). The game host is responsible for entering a sane value.

Also, the last player (who is typically the player who selected the tile
and must answer the daily double) won't be remembered/selected. The game
host has to do this manually.

If the player gives a wrong answer, the value is reset to the default
value of the answer.