Create print layout of a jeopardy board
This includes answers and questions and can be used as a cheat sheet by
the game host.

The print layout is made for a 5x5 board (adjust the CSS if you have a
different number of columns or rows). It will fit exactly on an A4 page.
Images will be scaled down, overflowing text gets hidden.

The footer contains the list of keyboard shortcuts that can be used
during the game.